requirements.txt
# Python preprocessing pipeline dependencies
# Install with: pip install -r requirements.txt

pandas>=2.0.0
scikit-surprise>=1.1.3
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.66.0
python-dotenv>=1.0.0
